libs/python/som/pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "cua-som"
version = "0.1.0"
description = "Computer Vision and OCR library for detecting and analyzing UI elements"
authors = [
    { name = "TryCua", email = "gh@trycua.com" }
]
dependencies = [
    "torch>=2.2.1",
    "torchvision>=0.17.1",
    "ultralytics>=8.1.28",
    "easyocr>=1.7.1",
    "numpy>=1.26.4",
    "pillow>=10.2.0",
    "setuptools>=75.8.1",
    "opencv-python-headless>=4.11.0.86",
    "matplotlib>=3.8.3",
    "huggingface-hub>=0.21.4",
    "supervision>=0.25.1",
    "typing-extensions>=4.9.0",
    "pydantic>=2.6.3"
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}
keywords = ["computer-vision", "ocr", "ui-analysis", "icon-detection"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition"
]

[project.urls]
Homepage = "https://github.com/trycua/cua"
Repository = "https://github.com/trycua/cua"
Documentation = "https://github.com/trycua/cua/tree/main/docs"

[tool.pdm]
distribution = true
package-type = "library"
src-layout = false

[tool.pdm.build]
includes = ["som/"]
source-includes = ["tests/", "README.md", "LICENSE"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "B", "I"]
fix = true

[tool.ruff.format]
docstring-code-format = true

[tool.mypy]
strict = true
python_version = "3.11"
ignore_missing_imports = true
disallow_untyped_defs = true
check_untyped_defs = true
warn_return_any = true
show_error_codes = true
warn_unused_ignores = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
